backend/services/chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = "chat application for channels"
authors = ["Kingsley Ocran <ocrankingsley@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from="." },
]

[tool.poetry.scripts]
chat = "scripts.runner:cli"


[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
daphne = "^4.0.0"
channels = "^4.0.0"
python-dotenv = "^1.0.0"
channels-redis = "^4.1.0"
gunicorn = "^20.1.0"
click = "^8.1.3"
django-prometheus = "^2.2.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-cors-headers = "^3.14.0"
django-elasticsearch-dsl = "^7.3"
opentelemetry-sdk = "^1.17.0"
opentelemetry-instrumentation-django = "^0.38b0"
opentelemetry-exporter-jaeger = "^1.17.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 90"


[tool.mypy]
# unfortunately cannot inject dependencies into super-linter
# so we have to ignore the imports as they will not have typehints
ignore_missing_imports = true

warn_unused_configs = true
disallow_any_generics = true
# disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
# disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
ignore_errors = true

[tool.black]
preview = true
